int->bigint sql datatype for github PR and workflow run ID.
Workflow run IDs don't fit in a 4 byte integer, and in PostgreSQL, the
integer type is 4 bytes (contrary to sqlite). This prevented exporting runs
with a workflow run ID set.

+++ service/models.py
@@ -6,7 +6,7 @@
 
 # Databases (SQLAlchemy) models.
 
-from sqlalchemy import (JSON, Boolean, Column, DateTime, Enum, ForeignKey,
+from sqlalchemy import (JSON, BigInteger, Boolean, Column, DateTime, Enum, ForeignKey,
                         Integer, String)
 from sqlalchemy.orm import relationship
 
@@ -34,7 +34,7 @@ class Run(Base):
     run_results = Column(JSON)
     source = Column(String, index=True)
     index_uid = Column(String, index=True)
-    github_pr = Column(Integer, index=True)
+    github_pr = Column(BigInteger, index=True)
     github_workflow_user = Column(String, index=True)
-    github_workflow_run_id = Column(Integer, index=True)
+    github_workflow_run_id = Column(BigInteger, index=True)
